nan fix

+++ src/decomposition/cp_decompose.py
@@ -121,10 +121,11 @@ def get_effective_weight(cp_seq: nn.Sequential) -> torch.Tensor:
     """
     pw_in = cp_seq[0]
     pw_out = cp_seq[-1]
-    # pw_in.weight: (R, C_in, 1, 1), pw_out.weight: (C_out, R, 1, 1)
-    w_in = pw_in.weight.squeeze(-1).squeeze(-1)   # (R, C_in)
-    w_out = pw_out.weight.squeeze(-1).squeeze(-1)  # (C_out, R)
-    return w_out @ w_in  # (C_out, C_in)
+    # Force fp32 — inside AMP autocast, matmul is cast to fp16 which overflows
+    # with SVD-initialized factors that have large entries
+    w_in = pw_in.weight.float().squeeze(-1).squeeze(-1)   # (R, C_in)
+    w_out = pw_out.weight.float().squeeze(-1).squeeze(-1)  # (C_out, R)
+    return w_out @ w_in  # (C_out, C_in) in fp32
 
 
 def compute_rank(in_channels: int, out_channels: int, rank_ratio: float) -> int:
